Changed find_end to make it easier to create search
